Start
The goal of the development is to help collect CAM data for CNC preparation from the 3D CAD model created during furniture design. By producing scarce 2D drawings. AutoCad 2025 in the form of plugins. Currently, blocks can be exported from the dwg file based on the original or foil. You can search for objects from the drawing space. The application has several layers, of which the 'CabinetBilder_AutoCad2025' project is currently under development.